refactoring, added strict mode
