+++ core/diskstorage.js
